Move WalletHolder into wallet module
