+++ app/server/src/main/scala/org/bitcoins/server/BitcoindRpcBackendUtil.scala
@@ -15,7 +15,7 @@ import org.bitcoins.dlc.wallet.DLCWallet
 import org.bitcoins.rpc.client.common.BitcoindRpcClient
 import org.bitcoins.rpc.client.v19.V19BlockFilterRpc
 import org.bitcoins.rpc.config.ZmqConfig
-import org.bitcoins.wallet.Wallet
+import org.bitcoins.wallet.{Wallet, WalletNotInitialized}
 import org.bitcoins.wallet.config.WalletAppConfig
 import org.bitcoins.zmq.ZMQSubscriber
